PBS-37 feature: Combine binlog_server with minimysql_server into a single executable (part 1)
https://perconadev.atlassian.net/browse/PBS-37

Refactored 'minimysql_server' code:
* Extracted frame-based network operations into
  'minimysql/network_io_operations.hpp'
* Extracted main MySQL protocol logic into
  'minimysql/network_service.hpp'
- Extracted sample binlog message data into
  'minimysql/sample_event_collection.hpp'

Where possible we now use 'boost/asio/ts/netfwd.hpp' for boost::asio forward
declarations.

Reworked the way how we suppress known false-positive warnings - we now
always use '#pragma GCC diagnostic ignored' for individual headers instead of
setting a '-Wno-error=...' flag on a whole compilation unit.
